Thu, Nov 3, 2022, 10:51 AM +00:00

+++ pset2/lecture/scores.c
@@ -23,5 +23,5 @@ float average(int length, int array[])
     {
         sum += array[i];
     }
-    return sum / length;
+    return sum / (float)length;
 }
